Refactor Graql query language library - Part 6 (#4731)
# Why is this PR needed?

This PR is a continuation of PR #4646, and incremental work for issue #4456 where you can find the full description of the goal. In summary, we are refactoring Graql to extract the Graql language library from Grakn's implementation of Graql, and we start with simplifying the Graql codebase first.

# What does the PR do?

This PR focuses on introducing a new language structure for Aggregate, Group and Group-Aggregate queries. 
1. Aggregate functions now come after `get` (fixes #4734, fixes #4119), and
2. Group query is no longer a subtype of an Aggregate query (fixes #4735).
    - We introduce `GroupQuery` separate from `AggregateQuery, and 
    - We introduce `GroupAggregateQuery` which is a subclass of `GroupQuery`
3. The execution/implementation of `GroupQuery` and `AggregateQuery` are also migrated to `QueryExecutor` and `AggregateExecutor` respectively, to get us closer to completing issue #4456.
    - We have a new/more efficient implementation for `mean`: we just consume the stream once and get the average value using `DoubleStream`
    - We have a new/more efficient implementation for `median`: we use a `MinHeap` and `MaxHeap` to find the median value. We consume the stream once and the runtime complexity should be `O(log(N))`

# Does it break backwards compatibility?

Yes. The syntax for Aggregate (and Group) queries need to change.

# List of future improvements not on this PR

Group query should be able to produce groups by more than 1 variable (issue #4740)