Refactor Transaction API - part 1 (#4976)
## What is the goal of this PR?

We are in the process of splitting `//client-java` to not transitively depend on `//server` (#4513), move to a separate repository (#4818), and split its second distribution for benchmarking (#4511). All to be done in PR #4926. However, we have a blocker:

To split client-java as described above, we need to split Grakn Transaction API (Session, Transaction, Keyspace interfaces)) into a separate module that does **not have any external dependency** that will be transitively brought into client-java. This is not is not the case right now, and the Concept API interfaces and classes are too complex that they need to be refactored first.

## What are the changes implemented in this PR?

- Replaced the use of `TransactionException` with `GraknClientException` in client-java
- Removed the use of AutoValue from Keyspace
- Renamed `Keyspace` to `KeyspaceImpl`, and introduced `Keyspace` interface in `//api` package
- Removed `throws` (javadoc) declaration in `Transaction` methods to `TransactionOLTP`
- Removed `@link` references in `TransactinoException` javadocs
- Replaced the constructor values of `Schema.MetaSchema` with `Graql.Token.Type`
- Replaced the use of `Schema.MetaSchema` in `Transaction` interface with `Graql.Token.Type`
- Replaced the use of `Schema.MetaSchema` in Graql queries with `Graql.Token.Type`