Deprioritise implicit labels (#5284)
## What is the goal of this PR?

We reinstate the previous behaviour when we assigned smaller priority to label fragments which correspond to implicit relations. This is due to the fact that in general and by default they correspond to edges (are non-reified).

## What are the changes implemented in this PR?

Differentiate between high- and low- priority starting point fragments - we define the label fragments to be low priority starting points.