Add console-deps deployment to repo as archive (#45)
## What is the goal of this PR?

Grakn core depends on console in order to create an "all" release package, however, it currently builds console from source (via bazel) which means we have two "releases" of console, one in this repo and one in the `grakn` repo, and they can have different sets of dependencies. This could cause any number of problems, not least being that tests run in this repo do not necessarily reflect the correctness of console in `grakn`. It also causes problems for us locally when using bazel to develop locally, because external workspaces imported in our own repo can leak into console unintended.

In order to fix this "artifact drift" we need to move to a "single true artifact" pattern. Currently, this artifact will only be `console-deps`, which does not include configuration (only `logback.xml` right now). At some point, we should move to the full `console` release being the distributed artifact and combined into the `grakn` "all" release.

## What are the changes implemented in this PR?

- Added `deploy-console-deps` distribution rule for `grakn` to depend on.
- Added CI/CD to deploy this.