Refactor WORKSPACE and add filegroup declaring CI targets (#68)
## What is the goal of this PR?
We have refactor the WORKSPACE file to make it clearer. The key change was to properly enclose declarations into appropriate sections. We also add a new toplevel BUILD file declaration for all targets used in CI and elsewhere to ensure these targets are correct at build time. We also bump the client-java commit.