Terminate console in the middle of the current answer calculation (#117)
## What is the goal of this PR?

We made the console more responsive by introducing the ability to terminate the answers stream in the middle of calculating the current answer.

## What are the changes implemented in this PR?

Instead of terminating the console only in between answers calculation, it can be canceled anytime during the answer computation. Although, the server will still run for some time in that case. 

close https://github.com/graknlabs/console/issues/93