Improve handling password input (#158)
## What is the goal of this PR?

We've updated Console such that password input can be supplied interactively and non-interactively.

This is done primarily for security purpose, interactive input will be the preferred one since it is not stored in the CLI history. The non-interactive mode is still relevant and therefore supported, particularly for use in automation scripts and test.

Additionally, we've refactored the code to improve the terminologies. We used the word "Command" everywhere to refer to several different things which cause confusion. We've now made the terminologies more precise:
  - "CLI Options" is how we refer to the options specified when starting console (eg., `--username admin`)
  - "REPL command" is how we refer to the Console commands such as 'database create'
  - We've renamed the execution mode to be clear. Console has three modes of execution: "REPL mode", "script mode" and "inline command mode". The first mode is the interactive mode that most users will use. The second and the third mode are non-interactive and is suitable for use in automation scripts or tests.

## What are the changes implemented in this PR?

- Add support for interactive password input
- Refactored various names in the code
  - Improve the clarity of the different execution mode
  - Fix the fact that we really like the word "command" which is used everywhere
- Make non-final variables final whenever possible