UX improvements to transaction commits and concept deletion (#178)
## What is the goal of this PR?

We've added an indicator for uncommitted changes within a transaction. While you are in a transaction that has uncommitted changes, an asterisk (*) will now be displayed in the prompt.

Also, concept deletion is now more transparent; the number of concepts being deleted (or updated) is printed to the console now. This is particularly helpful for catching when there is a bug in the match query that might cause it to delete no concepts, or the wrong concepts.

## What are the changes implemented in this PR?

- Indicate uncommitted changes
- Print the number of concepts that are being deleted (or updated)

We chose to not add a confirmation prompt prior to deletion just yet, this is due to Console architecture making that too difficult for now. This is (still!) tracked in:
- #166