Bump dependencies, updated factory CI instructions (#189)
## What is the goal of this PR?

We've bumped our dependencies and introduced updated instructions to Factory to facilitate the release of TypeDB 2.12.0. We've also updated our CI instructions to make use of Ubuntu 22.04 in line with our migration to Factory.

## What are the changes implemented in this PR?

We've fixed the Python version learned from the successful GitHub of typedb-common to a definite working version of Python for this job. 

We've also bumped our dependencies, fixing to the `2.12.0` tag for the upcoming release of TypeDB Console.