Cloud address translation (#241)
## Usage and product changes

We allow connection to the cloud servers using an address translation
mapping (cf. https://github.com/vaticle/typedb-driver/pull/624). This is
useful when the route from the user to the servers differs from the
route the servers are configured with (e.g. connection to public-facing
servers from an internal network).

Example usage: 
```bash
console \
    --cloud=typedb1.domain.com:1729=typedb.local:11729,typedb2.domain.com:1729=typedb.local:21729 \
    --username=<user> --password=<password>
```
or:
```bash
console \
    --cloud=typedb1.domain.com:1729=typedb.local:11729 \
    --cloud=typedb2.domain.com:1729=typedb.local:21729 \
    --username=<user> --password=<password>
```

Note: we currently require that the user provides translation for the
addresses of _all_ nodes in the Cloud deployment.

## Implementation

We also update the license headers and file from AGPL to MPL 2.0.