Rust console (#275)
## Usage and product changes

### Product changes

We re-implement TypeDB console in Rust, using a slightly more
sophisticated architecture. Most of the UX of the console remains the
same, with the exception of three usages:

**TLS enabled by default**

TypeDB console enables TLS connections by default. This further will
error if you try to connect to an HTTP endpoint, instead of an HTTPS
one. This change was required because username and password are now
required to connect to the any TypeDB 3.0 server.

To disable TLS (for example, for local development), use the
`--tls-disabled` flag when booting TypeDB console. Note that this will
send passwords in plaintext over the network.

**Replacement of `--core=<address>` and `--cloud=<address>` with
`--address=<address>`**

To connect to a TypeDB server (either a Core/CE 3.0, or a Cloud 3.0
deployment) use the new `--address=<address>` argument to the console.

**The command to open a transaction has changed**
```
>> transaction <db> <read|write|schema>
```
Has become:
```
>> transaction <read|write|schema> <db>
```
Note how the position of the database name has been swapped.

### Build

This repository is now supports dual Bazel/Cargo build systems.

To run TypeDB console directly from this repository, please see the
updated README!

### Artifacts

As before, Console will be published as platform-specific archives.
However, they now contain a single bash/bat file and the compiled
console binary. No more Java libs!

We also no longer deploy a `console-runner` artifact, since it was only
used to run the local `assembly` test. The assembly test is now written
in rust, using an unpublished, repository-local runner library that
drives both a Server and a Console in the test.

## Implementation

We throw out all the Java code in this repository, replacing it with
like-for-like functionality (though internal architecture is different).

The architecture is not ideal yet, however the rough structure is:

- `//src/repl`
This package holds a generic "Repl" + "Command" system, which can be
used to build custom REPLs, with layers of commands (`Subcommand`), leaf
commands (`CommandOption`), and fallback default commands
(`DefaultCommand`). The REPL also automatically builds `help` menus, and
allows typing `exit` to exit the program immediately.

Commands can also provide auto-completes and hidden input (if
requested).

- `//src/main.rs`
Contains the actual Console implementation, with dependent files (such
as `cli` for taking command-line inputs, `completions`

The two-layer REPL is implemented with a stack of REPL instances, held
in a `Context`. This is not particularly elegant, but can be extended in
the future. The stepping through the layers of REPL is currently driven
in the top-level `main`.

The TypeDB Rust-driver being used is the `async` version. However, the
REPL builder we use does not support async functions! As a result, we
have abstracted away the async driver behind a `Runtime`, which runs
tasks in a background `tokio` pool. This introduces some complexity when
we execute async functions, visible in `operations.rs`, but it is
managable. We have chosen this route to future console operations that
require the async driver, such as a bulk-loading functionality.