Multiline script paste and unified scripting format (#276)
## Usage and product changes

We refactor the Console to allow pasting a "script" into the current
REPL. This will be correctly interpreted and executed as a block of
separate commands. The inputs must look as they would when using the
console interactively:
```
database create tmp
transaction schema tmp
define entity person;

commit
```

With the empty newline, as in regular interactive console, being used to
mark the end of the query.

This is particularly useful for pasting fully self-contained examples
from TypeDB documentation into the console!

The format of `--file` scripts (now called `--script`) must follow this
same format, instead of using `\` to write multiline queries. We will
refer to this as the `TypeDB Console Script`, or `Scripted TypeQL` in
some contexts. We use `.tqls` for console scripts, instead of the usual
`.tql` which contains pure TypeQL.

## Implementation

We simplify the command definitions, and split the execution into two
phases:
1) match the command - including argument parsing
2) execute the command

We broadly try to parse matches using the index into the input, so that
we can split off ranges and pass them around to the next command parser.
This is particularly for the case because we want to split out multiple
commands from a multiline copy-paste.