Improve UX of Ctrl+C and fix bugs in command submission (#278)
## Usage and product changes

We an important bug:
1.
```
>> match
$x isa person; |
```
2. navigate cursor and enter newlines
```
>> match

| $x isa person;
```

Various things would break, including submitting the "script"
prematurely. Now, the block is only submitted when the sequence of
commands entered are a valid script.

Additionally, the UX if Ctr+C is improved: when hitting Ctrl+C on a
non-empty line, the line is retained, rather than cleared.

## Implementation
- hack around RustyLine to retain the previous line but not exit
- improve query/command submission logic