Support relative script and source commands (#293)
## Usage and product changes

We support using relative paths for the `--script` command (relative to
the current directory), as well as relative paths for the REPL `source`
command.

When `source` is invoked _from_ a script, the sourced file is
relativised to the script, rather than the current working directory.

## Implementation

* Store the console's working directory, and the script directory if a
script is active (scripts can't be nested, which enables this!)
* Relativize to either the script directory or the working directory as
when running with `--script` mode and `>> source <path>
` commands